refactor: 도메인 객체를 도메인 패키지로 이관한다
