+++ src/main/kotlin/introduce/domain/Introduce.kt
@@ -1,0 +1,5 @@
+package introduce.domain
+
+fun introduce(block: Person.() -> Unit): Person {
+    return Person().apply(block)
+}
