+++ src/main/kotlin/introduce/domain/Person.kt
@@ -1,0 +1,24 @@
+package introduce.domain
+
+class Person {
+    lateinit var name: String
+    lateinit var company: String
+    lateinit var skills: Skills
+    lateinit var languages: Languages
+
+    fun name(value: String) {
+        name = value
+    }
+
+    fun company(value: String) {
+        company = value
+    }
+
+    fun skills(block: Skills.() -> Unit) {
+        skills = Skills().apply(block)
+    }
+
+    fun languages(block: Languages.() -> Unit) {
+        languages = Languages().apply(block)
+    }
+}
